cleaning deps
