+++ rotala-client/src/bin/uist_client_test.rs
@@ -1,7 +1,8 @@
 use std::io::Result;
 
 use rotala::exchange::uist_v1::Order;
-use rotala_http::http::uist_v1::{Client, HttpClient};
+use rotala_client::client::uist_v1::HttpClient;
+use rotala_http::http::uist_v1::Client;
 
 #[tokio::main]
 async fn main() -> Result<()> {
