+++ rotala-client/src/client/mod.rs
@@ -1,0 +1,2 @@
+pub mod uist_v1;
+pub mod uist_v2;
