+++ rotala-client/src/lib.rs
@@ -53,6 +53,7 @@
 
 #[allow(unused)]
 pub mod broker;
+pub mod client;
 pub mod perf;
 pub mod schedule;
 pub mod strategy;
