+++ rotala-http/Cargo.toml
@@ -19,7 +19,6 @@ zip = "2.1.3"
 csv = "1.1.6"
 serde = { version = "1.0.193", features = ["derive"] }
 serde_json = "1.0.108"
-env_logger = "0.11.0"
 tokio = { version = "1.35.1", features = ["full"] }
 derive_more = { version = "1.0.0", features = ["full"] }
 anyhow = "1.0.86"
