+++ rotala-http/src/http/uist_v2.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 use rotala::exchange::uist_v2::{Order, Trade, UistV2};
 use rotala::input::athena::{Athena, DateBBO};
 
-type BacktestId = u64;
+pub type BacktestId = u64;
 
 pub struct BacktestState {
     pub id: BacktestId,
@@ -298,85 +298,6 @@ impl TestClient {
     pub fn single(name: &str, data: Athena) -> Self {
         Self {
             state: AppState::single(name, data),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct HttpClient {
-    pub path: String,
-    pub client: reqwest::Client,
-}
-
-impl Client for HttpClient {
-    async fn tick(&mut self, backtest_id: BacktestId) -> Result<TickResponse> {
-        Ok(self
-            .client
-            .get(self.path.clone() + format!("/backtest/{backtest_id}/tick").as_str())
-            .send()
-            .await?
-            .json::<TickResponse>()
-            .await?)
-    }
-
-    async fn insert_order(&mut self, order: Order, backtest_id: BacktestId) -> Result<()> {
-        let req = InsertOrderRequest { order };
-        Ok(self
-            .client
-            .post(self.path.clone() + format!("/backtest/{backtest_id}/insert_order").as_str())
-            .json(&req)
-            .send()
-            .await?
-            .json::<()>()
-            .await?)
-    }
-
-    async fn fetch_quotes(&mut self, backtest_id: BacktestId) -> Result<FetchQuotesResponse> {
-        Ok(self
-            .client
-            .get(self.path.clone() + format!("/backtest/{backtest_id}/fetch_quotes").as_str())
-            .send()
-            .await?
-            .json::<FetchQuotesResponse>()
-            .await?)
-    }
-
-    async fn init(&mut self, dataset_name: String) -> Result<InitResponse> {
-        Ok(self
-            .client
-            .get(self.path.clone() + format!("/init/{dataset_name}").as_str())
-            .send()
-            .await?
-            .json::<InitResponse>()
-            .await?)
-    }
-
-    async fn info(&mut self, backtest_id: BacktestId) -> Result<InfoResponse> {
-        Ok(self
-            .client
-            .get(self.path.clone() + format!("/backtest/{backtest_id}/info").as_str())
-            .send()
-            .await?
-            .json::<InfoResponse>()
-            .await?)
-    }
-
-    async fn now(&mut self, backtest_id: BacktestId) -> Result<NowResponse> {
-        Ok(self
-            .client
-            .get(self.path.clone() + format!("/backtest/{backtest_id}/now").as_str())
-            .send()
-            .await?
-            .json::<NowResponse>()
-            .await?)
-    }
-}
-
-impl HttpClient {
-    pub fn new(path: String) -> Self {
-        Self {
-            path,
-            client: reqwest::Client::new(),
         }
     }
 }
