+++ rotala/Cargo.toml
@@ -10,7 +10,6 @@ readme = "README.md"
 rust-version = "1.75"
 
 [dependencies]
-actix-web = "4"
 time = { version = "0.3.17", features = ["macros", "parsing"] }
 rand = "0.8.4"
 rand_distr = "0.4.1"
@@ -19,10 +18,7 @@ zip = "2.1.3"
 csv = "1.1.6"
 serde = { version = "1.0.193", features = ["derive"] }
 serde_json = "1.0.108"
-env_logger = "0.11.0"
 tokio = { version = "1.35.1", features = ["full"] }
-derive_more = { version = "1.0.0", features = ["full"] }
-anyhow = "1.0.86"
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["async_tokio"] }
